v1.0.3-d
- move comet joint orient and split joint from general to rig
- set attrs default now have att_name_list to have specific attr instead
whole attr visible in channel box
- add fix duplicate name, duplicate original mesh
- change function name from get_texture_pattern_files to glob_texture_files
- add MAYA_SCRIPT_PATH environment for FrMaya module to make onModelChange3ds.mel works
- add follicle uv and object position
